requirements.txt
django==5.0.1
drf-jwt==1.19.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-oauth-toolkit==2.3.0
pyjwt==2.8.0
python-dotenv==1.0.0